+++ CHANGELOG.md
@@ -53,6 +53,8 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#2714](https://github.com/Pycord-Development/pycord/pull/2714))
 - Added the ability to pass a `datetime.time` object to `format_dt`
   ([#2747](https://github.com/Pycord-Development/pycord/pull/2747))
+- Added `Guild.get_or_fetch()` shortcut method and better get_or_fetch
+  ([#2776](https://github.com/Pycord-Development/pycord/pull/2776))
 
 ### Fixed